ops: wire vercel cron for stats_timeseries generation
