+++ vercel.json
@@ -1,0 +1,16 @@
+{
+  "crons": [
+    {
+      "path": "/api/internal/cron/stats-timeseries?job=hourly",
+      "schedule": "0 * * * *"
+    },
+    {
+      "path": "/api/internal/cron/stats-timeseries?job=daily",
+      "schedule": "10 0 * * *"
+    },
+    {
+      "path": "/api/internal/cron/stats-timeseries?job=weekly",
+      "schedule": "20 0 * * 1"
+    }
+  ]
+}
